pyproject.toml
[tool.poetry]
name = "base"
version = "0.1.0"
description = ""
authors = ["bhstoller <bhs.stoller@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
cassio = "^0.1.8"
langchain = "^0.2.6"
langchain-community = "^0.2.6"
torch = "^2.3.1"
pypdf2 = "^3.0.1"
openai = "^1.35.7"
python-dotenv = "^1.0.1"
langchain-openai = "^0.1.13"
fitz = "^0.0.1.dev2"
frontend = "^0.0.3"
flask = "^3.0.3"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
